added tests

+++ main.py
@@ -26,6 +26,7 @@ def refresh(request:Request, response:Response):
     
     if is_valid:
         user = TokenHandler.get_user_bytoken(refresh_token)
+        if user is None: return {"status":"Failed", "message":"User not found"}
         TokenHandler.set_tokens(user, response)
         return {"status":"Ok"}
     else:
@@ -69,6 +70,7 @@ def login(input_user:loginUserModel, response:Response):
     hash = sha256(password.encode('utf-8')).hexdigest()    
 
     user = session.query(User).filter(User.username == input_user.username).first()
+    if user is None: return {"status":"Failed", "message":"User not found"}
     user_hash = user.password
 
     if user_hash == hash:
@@ -85,13 +87,6 @@ def logout(response:Response):
     except:
         return {"status": "Error"}
     
-
-@app.post("/test")
-def test(request:Request):
-    try:
-        AdminValidation(request)
-    except UserValidation.NotAuthorized as e:
-        return {"status":"NotAuthorized", "error":str(e.reason)}
 
 
 
@@ -119,7 +114,7 @@ def get_author(id:int, request:Request):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
 
-
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
     author = session.query(Author).filter(Author.id == id).first()
     return author
 
@@ -140,7 +135,9 @@ def update_author(request:Request, author_input:authorCreateModel, id:int):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
     
-    author = get_author(request, id)
+
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
+    author = get_author(id, request)
 
     try:
         author.name = author_input.name
@@ -148,6 +145,7 @@ def update_author(request:Request, author_input:authorCreateModel, id:int):
         author.birth_date = author_input.birth_date
         session.commit()
     except:
+        session.rollback()
         return {"status":"Error"}
 
     return {"status": "Ok"}
@@ -159,13 +157,15 @@ def delete_author(request:Request, id:int):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
 
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
 
     try:
-        author = get_author(request, id)
+        author = get_author(id, request)
+        if author is None: return {"status":"Failed", "message":"Author not found"}
         session.delete(author)
         session.commit()
     except:
-        {"status":"Failed"}
+        return {"status":"Failed"}
     
     return {"status": "Ok"}
 
@@ -178,6 +178,7 @@ def create_book(request:Request, book_input:bookCreateModel):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
     
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
     author = session.query(Author).filter(Author.id == book_input.author_id).first()
     if author is None: return {"status":"Error","message":f"Author with author_id = {book_input.author_id} doesnt exist."}
 
@@ -198,7 +199,7 @@ def get_book(request:Request, id:int):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
 
-
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
     book = session.query(Book).filter(Book.id == id).first()
     return book
 
@@ -219,6 +220,7 @@ def update_book(request:Request, book_input:bookCreateModel, id:int):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
     
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
     author = session.query(Author).filter(Author.id == book_input.author_id).first()
     if author is None: return {"status":"Error","message":f"Author with author_id = {book_input.author_id} doesnt exist."}
 
@@ -244,6 +246,7 @@ def delete_book(request:Request, id:int):
     except UserValidation.NotAuthorized as e:
         return {"status":"NotAuthorized", "error":str(e.reason)}
 
+    if id not in range(-2_147_483_647, 2_147_483_647): return {"status":"Error", "message":"value out of int range"}
 
     try:
         book = get_book(request, id)
